+++ packages/components/rollup.config.js
@@ -25,58 +25,8 @@ export default [
     dtsExternal: [/\.scss$/],
     copy: [["client/styles", "client"]],
   }),
-  ...rollupTypescript("client/components/ArticleInfo", {
-    external: [
-      "@mr-hope/vuepress-shared/lib/client",
-      "@vuepress/client",
-      "balloon-css/balloon.css",
-      "vue",
-      "vue-router",
-      /\.scss$/,
-    ],
-    dtsExternal: ["balloon-css/balloon.css", /\.scss$/],
-  }),
   ...rollupTypescript("client/components/Badge", {
     external: ["vue", /\.scss$/],
-    dtsExternal: [/\.scss$/],
-  }),
-  ...rollupTypescript("client/components/BreadCrumb", {
-    external: [
-      "@mr-hope/vuepress-shared/lib/client",
-      "@vuepress/client",
-      "@vuepress/shared",
-      "vue",
-      "vue-router",
-      /\.scss$/,
-    ],
-    dtsExternal: [/\.scss$/],
-  }),
-  ...rollupTypescript("client/components/FullScreen", {
-    external: [
-      "@mr-hope/vuepress-shared/lib/client",
-      "@vueuse/core",
-      "vue",
-      /\.scss$/,
-    ],
-    dtsExternal: [/\.scss$/],
-  }),
-  ...rollupTypescript("client/components/Pagination", {
-    external: [
-      "@mr-hope/vuepress-shared/lib/client",
-      "vue",
-      "vue-router",
-      /\.scss$/,
-    ],
-    dtsExternal: [/\.scss$/],
-  }),
-  ...rollupTypescript("client/components/TOC", {
-    external: [
-      "@mr-hope/vuepress-shared/lib/client",
-      "@vuepress/client",
-      "vue",
-      "vue-router",
-      /\.scss$/,
-    ],
     dtsExternal: [/\.scss$/],
   }),
   ...rollupTypescript("client/root-components/BackToTop", {